requirements.txt
configparser
flask
redis
request
gunicorn
RPi.GPIO
ds18b20
